Fix: Improve histogram binning and x_fit range (master)
Improved the histogram binning by using a more appropriate
calculation.

Modified the range of `x_fit` used for plotting the
fitted double Gaussian curve. The new range is from
`0.25 * total_value` to `0.75 * total_value`, which
provides a better visualization of the fit, especially
when the distribution is not centered around the middle
of the histogram.

Signed-off-by: Aleksander Augustyn <aleksander.augustyn@ncbj.gov.pl>